pyproject.toml
[tool.poetry]
name = "gerenciador escolar"
version = "1.0.0"
authors = ["vini-devizin <viniciusdaniel.br@gmail.com>"]
readme = "README.md"

packages = [
	{ include = "code" }
]

[tool.poetry.dependencies]
python = "^3.10"
psycopg2 = "2.9.10"
python-dotenv = "1.0.1"
tabulate = "0.9.0"
pyjwt = "^2.8.0"
bcrypt = "4.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

